Try a simpler API
Let's drop the Generics and TypeClasses for the moment, and just create
a set of functions that help you compose Elm types for Haskell types.

This first draft deals with what I think is most of the hard type
problems. It demonstrates elm types can be created for primitves and
arbitrary sums and products. Both the range of types expressed and the
API will need improvement, but it's a good first step.