Add `HasElm` typeclass
Hedgehog.Gen shows how manually building 'instances' for types doesn't
need to be terrible. Unfortunately in the case of `ToElm`, building
instances for sum or product types is not nice, because it requires a
whole bunch of `Vinyl` manipulation.

I don't see how to bring this out. In the end, Elm supports arbitrary
sum and product types. I don't want users to manually derive JSON
encoders and decoders with their Elm types, because there's lots of
scope for errors here (nobody guarantees the Haskell encoders and
decoders passed match the ones the library will generate for Elm). Also
I'd like the option to switch JSON serialization for something else at
some point.

This means users will need to define transformations to and from a
Haskell type that can express arbitrary sums and products. There's no
support for this in the standard library, so we need something like
Vinyl. But that's a tough library to learn to use.

Defining a class with instances for Vinyl (Co)Rec's might reduce the
amount of Vinyl code people will need to write. If that's not enough,
maybe a `Generic` based approach needs to be supported for Haskell
products and sums, so Vinyl can be avoided entirely.

Either I way I plan to expose the manual builder API too, so people can
still override these automatic derivations with their own logic.