Allow CoRec's with fields to take a functor
This is a step towards being able to use our machinery to expresses
coders for sums of products.

At the moment we can express a flat product or a flat sum. To be able to
represent all elm types, we'll also need to be able products nested in
sums. For example, the following type nests products in sums:

  type IntString = OnlyInt Int | IntAndString Int String

The `IntAndString` branch here contains a product of two types.

In terms of the `Vinyl` types we're defining, this requires us will
require us to nest a `Rec` in a `CoRec`. There's currently no place for
that, so we make place by adding an `f` type parameter where we are
going to nest our `Rec`.

For the current applications `f` is made equal to `Identity`. This extra
wrapping makes manually building simple union types uglier then before,
but the interface we have for that now was to ugly to serve as the final
UI for the library anyway. Complicating the interface to allow us to
write a simpler one going forward seems worth it.