Start extracting a custom version of `FieldRec`
Up to this point we've hacked around Vinyl's bundled `ElField` and
`FieldRec` types, and this has borne results. Now that we're looking to
build on the work we did, notably to add records into the custom records
we had, it's time to turns the hacks in a nice abstraction that can be
worked with more easily.

This abstraction is an alernative to `FieldRec` that better fits the
purpose. We'll write helper functions that can act on this type
directly. The `Field` type at its basis is a generalization of
`ElField`, and so we can remove `ElField` from our codebase entirely so
there's one less concept to know about.

This commit gets us started by setting up some types synonyms. We
haven't cleaned up much of the logic yet to make use of this new way of
working.