Remove `ToElm` in favor of `IsElmType`
`ToElm` described the how: Given a type, how to encode it on the Haskell
side? How to generate an elm type for it? Soon we'd need to add how to
generate Elm encoders and decoders for it.

`IsElmType` describes what: what Elm Types exist? From that information,
we can derive all the hows.

Having this core set of Haskell types that represent Elm types, the job
of adding support for other Haskell types not in the core set becomes
easier. We will only need to provide transformation functions in both
directions between our custom Haskell type and one of the types in the
set.

We use a GADT to describe Elm Types rather than type classes to
hopefully make our live easier in a bit when we start adding
Generics-derivations for Haskell records and sums. The hope is that this
way we won't need to proof to the compiler that a `HasElm` constraint
hold for a certain composite type, given it holds for its component
types. The GADT should allow us to stay on the value level.

This is why it's not immediately problematic that the `HasElm` code is
losing the ability to derive instances for Vinyl Records and SOPs. That
was never the goal, so it's okay to loose this.