Wrap record in constructor and data type
The `MC` (constructor) and `MD` (data type) wrappers of the Generic
primitives we passed through before, resulting in an Elm record alias
being produced. Now we apply those wrappings resulting in a custom type
wrapper around the record in Elm.

Two things are left to be done to complete the Generics work: we need to
handle multiple parameters and multiple constructors.