Remove Dhall code
The goal of this repo was always to generate Elm types from Haskell
types. Dhall was an interesting route to that, but I realized there's a
problem with it: When you turn a Haskell type into its Dhall
representation you loose the name of the type. We'll not be able to
reconstruct that on the Elm side. That's an unacceptable loss in
ergonomics.

Going to try using Generics instead. I tried before, but with a more
complicated Elm representation. The simpler representation I created for
working with Dhall should offer a fresh start.