Remove dependency on HashMap
Having just removed the use of `HashMap` in the representation of Elm
sum types (because we wanted to make empty sum types unrepresentable), I
started wondering if we should get rid of it for records as well. The
advantages would be being able to remove a dependency from the project.

The disadvantage would be slower field access, but this it turns out is
something we don't actually need in the code.

So we remove HashMap from the project entirely.