Dis-ambiguitate wire type
The wire type allowed similar values to be expressed in multiple ways. A
primitive type could be expressed as just that type, or as a
single-valued product containing the type. That ambiguity would need to
be handled in the Elm and Haskell code in the same way.

By ensuring primitives are always in a product we remove the ambiguity.