Simple Haskell encoding/decoding
I don't like using the default Aeson encoders and decoders. It takes a
dependency on the encoding/decoding style of the aeson library and it
results in very large payloads. I've experimented with other approaches
but it's eating up more and more time, so I'm going to go with those
encoders/decoders for now and work on something else for a bit. Maybe
I'll revisit them later.