Wrap type construction in a functor
We're currently not handling recursive types. These will result in the
code looping indefinitely while it keeps walking through the recursion.

To break the loop we'll have to flatten our types and be able to detect
during iteration which types we've already seen. By wrapping the
construction process of the type in a functor we'll front-load the work
required to get our generics code compiling.