Bring wire value in line with wire type
In the wire type each constructor of a sum has a single subtype. That
subtype can be a product if it represents a constructor with multiple
parameters.

In the wire value each constructor was defined as a product of subtypes.
That was in addition to there also existing a stand-alone product.

We changes the wire value to also contain a single subvalue that can be
a product.