Add custom encoding and decoding logic
We were using Aeson-derived encoders and decoders for the `Wire.Value`
type. Those are cheap to generate, but very noisy. Every primitive value
(ints, floats, strings) are wrapped in a JSON object with a string
describing the constructor used, to allow decoding logic to know
whether, for example, a particular JSON number is meant to be an `Int`
or a `Float`.

What we'd like and what this commit implements are encoders and decoders
that drop some of this wrapping. The disadvantage is that we will no
longer be able to implement `Aeson` instances for the `Wire.Value` type,
because the encoded data does not in itself have enough context to know
how to decode it. Using that same example: if we have a JSON number
somewhere we can't know whether that is intended to represent a `Float`
or an `Int`.

That context is available however in the `Wire.Type_` definition of the
type. And so by passing the value and the type we mean to get out of it
we can decode that value.