Don't use `Text` for various kinds of names
Use separate types for names of types, constructors, and fields. This
better communicates intent in various pieces of code and should make it
harder to accidentally mix up names.

The wire type names are currently used in the Elm module as well. This
is likely a temporary state of affairs. We'll need to write some logic
to turn the wire type names into Elm names that are guaranteed to be
unique.