Distinguish between products and tuples
So far the wire type didn't distinguish between product and tuple types.
Wire type values didn't have any record field names. This seemed simpler
when encoding and decoding.

The downside of not encoding record field names is that it becomes way
easier to break backwards compatibility. Where previously removing a
field from a service would still allow other services to call the
service with the extra field (it would simply be ignored), now such a
change is breaking because we don't know which field got removed.

To improve backwards compatibility we start encoding field names, but
this requires us to treat records and tuples separately. Tuples do not
have any field names, and so where we are entirely reliant on their
order.