Add support for `Map` and `HashMap` to wire format
These will both be encoded as lists of tuples. It would maybe be nicer
to have a separate type for maps, but that will require more work on the
Elm side to figure out which of those maps can and which cannot be
represented with `Dict`, depending on the type of the keys.

For the moment we'll just default to list of tuples. That always works,
and if it's a format from which it's easy enough to generate
dictionaries yourself if you so choose.