Include type parameters in type name
Previously we were using the qualified type constructor name as the full
name of the type. This did not include any of the type's parameters. As
a result `Maybe Int` and `Maybe Text` would have the same type name. And
because that type name is used by us to index types in a map, a type
that included multiple fields wrapped in Maybe would only see the type
of one of those fields overwriting the others. We don't have an instance
`Wire a => Wire (Maybe a)` yet, but when we do this will become a
problem.

GHC.Generics does not give us access to the names of parameters. We
might be able to set something up for the first parameter by defining
`Generic1` instances, but we want a solution that works for types with
arbitrary amounts of parameters. The `Type.Reflection` module can give
us information about the type parameters.

For the moment we will not render the type parameters in our Elm code,
because the parameters reference Haskell types, not Elm types.
That's potentially going to result in duplicate type names. We already
have sources of duplicate Elm type names though (for example, elm types
generated for different Haskell types with the same name but in
different modules). This is an open issue to look into.