Restrict allowed parameter types
In principle Servant allows as parameter types anything that implements
the `FromHTTPApiData` / `ToHTTPApiData` type classes. The problem with
this for code generation is that we cannot introspect the
implementations of these instances for a given type. As a result we're
not able to write matching implementations in Elm.

To protect against this we severly limit the types allowed for
parameters, to a couple of primitive types and wrappers of those
primitive types. For the primitive types we know the `FromHTTPApiData`
and `ToHTTPApiData` implementations used because the instances are
bundled with Servant. This means we can be pretty sure our generated Elm
code will match these implementations.

For the wrapper types users are theoretically able to implement
non-standard instances. If they do Elm code generation will not match
these custom instances. I don't see a way around this without forbidding
wrapper types all together.