Add helper modules for a roundtrip test
The servant API exposes endpoints for getting and posting a Haskell type
(currently a wrapped Int). Then we have an Elm app that gets data from
one endpoint and immediately pushes it out again on the other.

Now we have to write a test to run these against each other.