Unwrap single params in Encoding
We're already doing this on the Elm side.

I'm wondering if we should remove all these optimizations. It makes it
much easier to implement matching encoding/decoding logic in multiple
languages, at the cost of larger payloads. The bigger payloads aren't
nice, but Gzip and/or support for a binary protocol would take care of
that easily.